requirements.txt
anthropic==0.34.0
yfinance==0.2.36
aiohttp==3.9.1
python-dotenv==1.0.0
